refactor(crouton-printing): move ESC/POS engine and driver registry from crouton-sales (#327)
👤 For humans
Moves the "turn a receipt into printer bytes / pick the right driver" code
out of the sales package into a new shared @fyit/crouton-printing package. No
behaviour change — the same receipts print byte-for-byte. The printing brain
now lives where any package can reuse it.

🤖 For agents
Lift-and-shift (epic #325, issue #327):
- New package packages/crouton-printing with the domain-neutral engine:
  - server/utils/receipt-formatter.ts — ESC/POS CP858 byte builder
    (formatReceipt), HTML sibling renderTicketHtml, formatTestReceipt, the
    canonical ReceiptData/ReceiptItem/ReceiptSettings types. Verbatim copy.
  - server/utils/print-queue-service.ts — encodeTicket, PRINT_STATUS,
    receiptCurrencySymbol, routing (generatePrintJobsForOrder etc.).
  - server/utils/driver-registry.ts — NEW explicit output-driver registry
    behind encodeTicket(payload, driver): network-escpos (base64 ESC/POS) +
    browser-print (JSON) registered; serial/usb slot in via registerDriver.
    generatePrintJobsForOrder routes to any registered driver (isDriverRegistered),
    preserving the old `['network-escpos','browser-print']` behaviour exactly.
  - schemas/printers.json, schemas/printQueues.json — copied here.
  - test/receipt-formatter.test.ts — byte-for-byte base64 snapshot (customer
    receipt + kitchen ticket, accents/€/options on a fixed instant+timezone)
    plus encodeTicket driver-routing assertions.
- crouton-sales keeps compiling via thin re-export shims at the old paths
  (server/utils/{receipt-formatter,print-queue-service}.ts), surface identical
  to pre-move; slated for removal in #325/5. Adds @fyit/crouton-printing dep.
- Sales-specific glue (generate-print-queues.ts) stays in crouton-sales.
- Sales crouton.manifest.ts still imports its own schemas/*.json (unchanged),
  so sales schema copies remain for now.
- No DB or endpoint changes. pnpm -r --filter './apps/*' typecheck green.
