+++ packages/crouton-printing/schemas/printers.json
@@ -1,0 +1,12 @@
+{
+  "id": { "type": "uuid", "meta": { "primaryKey": true } },
+  "eventId": { "type": "uuid", "refTarget": "events", "meta": { "required": true, "label": "Event" } },
+  "locationId": { "type": "uuid", "refTarget": "locations", "meta": { "label": "Location", "helpText": "Routing key for kitchen printers only — receipt printers ignore it (PrinterForm requires it for type kitchen)" } },
+  "title": { "type": "string", "meta": { "required": true, "label": "Printer Name", "maxLength": 100 } },
+  "ipAddress": { "type": "string", "meta": { "required": true, "label": "IP Address" } },
+  "port": { "type": "integer", "meta": { "label": "Port", "default": 9100 } },
+  "type": { "type": "string", "meta": { "label": "Printer Type", "default": "kitchen", "helpText": "kitchen = per-location tickets, receipt = combined customer receipt with footer text" } },
+  "status": { "type": "integer", "meta": { "label": "Status", "default": 0, "helpText": "0=ready, 1=printing, 9=error" } },
+  "showPrices": { "type": "boolean", "meta": { "label": "Show Prices on Receipt", "default": true } },
+  "isActive": { "type": "boolean", "meta": { "label": "Active", "default": true } }
+}
